requirements.txt
# AWS SDK for Claude API access
boto3>=1.28.0

# Data analysis and visualization
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.12.0

# PDF processing
markitdown[pdf]>=0.1.1

# Environment variables
python-dotenv>=1.0.0

# File handling
pillow>=9.0.0
tqdm>=4.64.0

# Legacy dependencies (if needed)
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# AWS Bedrock specific
anthropic>=0.5.0